Update Passenger.java
Imported Jakarta for use of Entity, Id, GeneratedValue and GenerationType annotations. Overall purpose is to:

- @Entity indicates that the given class is an entity to be used in a database
@Id marks the Primary Key (PK) in said entity class
@GeneratedValue will specify how the PK is generated, i.e generated automatically
@ManyToOne, OneToMany defines relationships between the entities (just realized I forgot some capital letters on the other commits)

Besides that, I added the private objects and haven't touched the mutators/accessors, which are next. (this commit desc. will be C&P'd for the other classes so you can safely ignore the other commits)